Recreate and move Gradle bare-minimum example to dedicated directory
