--- nflow-examples/spring-boot/bare-minimum/build.gradle
@@ -1,22 +1,0 @@
-plugins {
-  id "com.github.ben-manes.versions" version "0.27.0"
-  id 'org.springframework.boot' version '2.2.1.RELEASE'
-  id 'io.spring.dependency-management' version '1.0.8.RELEASE'
-}
-
-apply plugin: 'java'
-apply plugin: 'eclipse'
-
-version = '0.0.1-SNAPSHOT'
-sourceCompatibility = 1.8
-
-repositories {
-  mavenCentral()
-}
-
-dependencies {
-  implementation 'org.springframework.boot:spring-boot-starter-jdbc'
-  implementation 'io.nflow:nflow-engine:7.0.0'
-  runtimeOnly 'com.h2database:h2'
-  testImplementation 'org.springframework.boot:spring-boot-starter-test'
-}
