+++ nflow-examples/spring-boot/bare-minimum/gradle/settings.gradle
@@ -1,0 +1,1 @@
+rootProject.name = 'demo'
