+++ nflow-examples/spring-boot/bare-minimum/gradle/src/main/resources/application.properties
@@ -1,0 +1,2 @@
+spring.profiles.active=nflow.db.h2
+logging.level.root=WARN
