+++ nflow-examples/spring-boot/bare-minimum/gradle/src/test/java/io/nflow/springboot/bareminimum/gradle/DemoApplicationTests.java
@@ -1,0 +1,13 @@
+package io.nflow.springboot.bareminimum.gradle;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class DemoApplicationTests {
+
+	@Test
+	void contextLoads() {
+	}
+
+}
