--- nflow-examples/spring-boot/bare-minimum/gradle/wrapper/gradle-wrapper.properties
@@ -1,6 +1,0 @@
-distributionBase=GRADLE_USER_HOME
-distributionPath=wrapper/dists
-# upgrading to 5.6 or newer breaks eclipse test classpath
-distributionUrl=https\://services.gradle.org/distributions/gradle-5.5.1-bin.zip
-zipStoreBase=GRADLE_USER_HOME
-zipStorePath=wrapper/dists
