--- nflow-examples/spring-boot/bare-minimum/src/main/resources/application.properties
@@ -1,3 +1,0 @@
-spring.profiles.active=nflow.db.h2
-logging.level.root=WARN
-nflow.db.h2.url=jdbc:h2:mem:test;TRACE_LEVEL_FILE=4;DB_CLOSE_ON_EXIT=FALSE
