--- nflow-examples/spring-boot/bare-minimum/src/test/java/io/nflow/springboot/bareminimum/SpringBootBareMinimumApplicationTests.java
@@ -1,17 +1,0 @@
-package io.nflow.springboot.bareminimum;
-
-import org.junit.Test;
-import org.junit.runner.RunWith;
-import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.test.context.junit4.SpringRunner;
-
-@RunWith(SpringRunner.class)
-@SpringBootTest
-public class SpringBootBareMinimumApplicationTests {
-
-  @Test
-  public void applicationStarts() {
-    // just check that the application starts
-  }
-
-}
